# allocator optimization: avoid tmp obj's allocation and gc