--- a/alloc_test.go
+++ b/alloc_test.go
@@ -32,25 +32,25 @@ func TestAllocGet(t *testing.T) {
 	if alloc.Get(0) != nil {
 		t.Fatal(0)
 	}
-	if len(alloc.Get(1)) != 1 {
+	if len(*alloc.Get(1)) != 1 {
 		t.Fatal(1)
 	}
-	if len(alloc.Get(2)) != 2 {
+	if len(*alloc.Get(2)) != 2 {
 		t.Fatal(2)
 	}
-	if len(alloc.Get(3)) != 3 || cap(alloc.Get(3)) != 4 {
+	if len(*alloc.Get(3)) != 3 || cap(*alloc.Get(3)) != 4 {
 		t.Fatal(3)
 	}
-	if len(alloc.Get(4)) != 4 {
+	if len(*alloc.Get(4)) != 4 {
 		t.Fatal(4)
 	}
-	if len(alloc.Get(1023)) != 1023 || cap(alloc.Get(1023)) != 1024 {
+	if len(*alloc.Get(1023)) != 1023 || cap(*alloc.Get(1023)) != 1024 {
 		t.Fatal(1023)
 	}
-	if len(alloc.Get(1024)) != 1024 {
+	if len(*alloc.Get(1024)) != 1024 {
 		t.Fatal(1024)
 	}
-	if len(alloc.Get(65536)) != 65536 {
+	if len(*alloc.Get(65536)) != 65536 {
 		t.Fatal(65536)
 	}
 	if alloc.Get(65537) != nil {
@@ -63,19 +63,24 @@ func TestAllocPut(t *testing.T) {
 	if err := alloc.Put(nil); err == nil {
 		t.Fatal("put nil misbehavior")
 	}
-	if err := alloc.Put(make([]byte, 3, 3)); err == nil {
+	b := make([]byte, 3)
+	if err := alloc.Put(&b); err == nil {
 		t.Fatal("put elem:3 []bytes misbehavior")
 	}
-	if err := alloc.Put(make([]byte, 4, 4)); err != nil {
+	b = make([]byte, 4)
+	if err := alloc.Put(&b); err != nil {
 		t.Fatal("put elem:4 []bytes misbehavior")
 	}
-	if err := alloc.Put(make([]byte, 1023, 1024)); err != nil {
+	b = make([]byte, 1023, 1024)
+	if err := alloc.Put(&b); err != nil {
 		t.Fatal("put elem:1024 []bytes misbehavior")
 	}
-	if err := alloc.Put(make([]byte, 65536, 65536)); err != nil {
+	b = make([]byte, 65536)
+	if err := alloc.Put(&b); err != nil {
 		t.Fatal("put elem:65536 []bytes misbehavior")
 	}
-	if err := alloc.Put(make([]byte, 65537, 65537)); err == nil {
+	b = make([]byte, 65537)
+	if err := alloc.Put(&b); err == nil {
 		t.Fatal("put elem:65537 []bytes misbehavior")
 	}
 }
@@ -85,7 +90,7 @@ func TestAllocPutThenGet(t *testing.T) {
 	data := alloc.Get(4)
 	alloc.Put(data)
 	newData := alloc.Get(4)
-	if cap(data) != cap(newData) {
+	if cap(*data) != cap(*newData) {
 		t.Fatal("different cap while alloc.Get()")
 	}
 }
@@ -93,5 +98,12 @@ func TestAllocPutThenGet(t *testing.T) {
 func BenchmarkMSB(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		msb(rand.Int())
+	}
+}
+
+func BenchmarkAlloc(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		pbuf := defaultAllocator.Get(i % 65536)
+		defaultAllocator.Put(pbuf)
 	}
 }