pool处理的buffer类型由 `[]byte` 改为 `*[]byte`，避免 `[]byte` 作为函数参数时的临时对象拷贝和不必要分配、gc压力。
```golang
func get() ([]byte, uintptr) {
	b := make([]byte, 10)
	return b, uintptr(unsafe.Pointer(&b))
}
func main() {
	b, p1 := get()
	p2 := uintptr(unsafe.Pointer(&b))
	println("p1:", p1, "p2:", p2, p1 == p2)
}
```
```sh
p1: 1374390779568 p2: 1374390779664 false
```

benchmark代码:
```go
func BenchmarkAlloc(b *testing.B) {
	for i := 0; i < b.N; i++ {
		b := defaultAllocator.Get(i % 65536)
		defaultAllocator.Put(b)
	}
}

```

修改前使用 `[]byte`：
```sh
goos: darwin
goarch: arm64
pkg: github.com/xtaci/smux
cpu: Apple M4 Pro
BenchmarkAlloc-14    	62714854	        19.06 ns/op	      24 B/op	       1 allocs/op
PASS
ok  	github.com/xtaci/smux	2.349s
```

修改为 `*[]byte` 后：
```sh
goos: darwin
goarch: arm64
pkg: github.com/xtaci/smux
cpu: Apple M4 Pro
BenchmarkAlloc-14    	155811184	         7.626 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/xtaci/smux	2.376s
```

前阵子有人提到这个优化，自己的大改了一轮，相关讨论：
https://github.com/lesismal/nbio/issues/454